work around Ruby bug empty files in YAML.load_file
Ruby 1.9.3p194 includes a change to Psych to support BOM markers. Unfortunately, that change means that Psych will sometimes throw an exception when you try to call YAML.load_file on a file with zero bytes in it. This commit includes a workaround so that we don't call YAML.load_file if the config file is empty.