rework builds to reduce false build failures
  - stop testing 1.9.2 against 1.5.3, 1.6.2, 1.7.2
  - mark the build as passing even if 1.9.2 builds fail

MRI 1.9.2 is really. fucking. slow.

It's so slow, in fact, that running the entire Bundler test suite frequently 
takes more than 25 minutes, at which point Travis times out the build. The 
test suite itself is not actually that slow, running in 8 minutes on 1.8.7 
and 7 minutes on 2.0.0-preview2. To work around this, we no longer build the 
older versions of Rubygems against 1.9.2, and we allow failures on 1.9.2 
without marking the overall build as failed.