lockfile source parity
the lockfile parser was creating and then providing a single rubygems
source, with multiple remotes. that doesn't work now that we have
multiple remotes, so I updated it to create source objects for all
of the remotes that are in the lockfile.

this doesn't solve the problem of source affinity at all, and at best
just punts on it, but we definitely need the lockfile parser to return
multiple rubygems sources instead of just one.