Revert "reinstalls gem if gem dir is missing but spec exists"
It turns out that this has some pretty extensive repurcussions, and
makes it impossible to use Bundler in some specific circumstances, like
with gems that ship with Ruby, and with gems that are installed by distro
packages. It also breaks the `bundle pack` command, because the gems
that are cached by the pack command are not found by this check.

See also:

https://github.com/bundler/bundler/issues/2854
https://github.com/bundler/bundler/issues/2818
https://github.com/bundler/bundler/issues/2780

This reverts commit 27bd51d8dbc38ae639f59430988c3124faf00db7.