Ensure Gemspec.lock and Gemfile match in multi-platform cache spec
The activerecord 2.3.3 gem needs activesupport 2.3.2, which was not
included in the Gemfile.lock or cache. This caused bundle update to
fail. This change simplifies the set of needed gems, thereby returning
to an older version of this spec. By adding the mri platforms to the
block for 'rack' we keep the check for allowed platforms.