Hardcode the list of options for `bundle install --trust-policy`
The reasoning behind this is that OpenSSL gets loaded up in a situation
where we just want to show the help banner for the install command.

Seen as the security policies on the rubygems repo haven't been changed
in over two years, it's safe to just hardcode the values, thus generally
speeding up Bundler commands, most notably in JRuby.

Closes #2950

Conflicts:
	lib/bundler/cli.rb