Revert "Handle s3 source"
Unfortunately, AWS S3 credentials may contain characters that are not
valid in a URI. That means that the credentials cannot be provided in
the URI "basic auth" format, which is the only format Bundler currently
supports. As a result, it seems that this feature is broken both in
Rubygems and here in Bundler. Attempts to fix the problem[1] have
stalled out, and no one on the Bundler core team uses S3 sources, so
we've decided to not release support for S3 sources as part of the
Bundler gem itself.

We're working on a plugin system that will allow S3 support to be
implemented and updated without requiring new versions of Bundler to be
released. In the meantime, if this support is very important to allow
you to continue to work, you can use the the 1.8.0.svn gem as a
workaround.

[1]: https://github.com/bundler/bundler/pull/3345

This reverts commit ec38f8a2c7b3bee861426be068f9a8bf2c0e4e23.
This reverts commit 300fc69447a8db7fc92227802e84178ed4035797.
This reverts commit 485c4469315333f12efe2d6ee8c7d52b1ef2b6b5.
This reverts commit 5d20e16875f23116353ff8b8725da38f2bdda151.