Refactoring config.rb
My main goal was to show the structure of the run method:
1. Guard Clauses
2. Message to user
3. Save new settings

and to separate the message texts from the rest of the logic.
So the class and methods are easier to understand read.

While doing that I reduced the complexity of the run method by
extracting methods. This should improve the Code Climate rating of
Bundler::CLI::Config substantially.