Auto merge of #4611 - BrianHawley:fix_20160524_require_env, r=segiddins
Fix the combination of gem :require and env or install_if

If you guard the installation of a gem with `env` or `install_if`, and that gem declaration has a `:require` option specified, it tried to require the gem even though it hadn't been installed. It looks like a spot was missed when the `env` command was added in the first place.