Detect changes in path specs by comparing locked specs to source specs
Previously, we'd wrongly conclude there are no path changes when there are
changes. We'd parse the Gemfile.lock and compare the source.specs, the
gemspecs currently in the path, to what's in the Gemfile.lock's locked.specs.
Unfortunately, locked.specs for Path sources creates an Index with the specs
from the filesystem and NOT what we already parsed from the Gemfile.lock!  In
other words, we compare the filesystems specs for a path to itself and always
conclude "No changes detected!"

Instead, we build an index with specs for the source we want from the
already parsed Gemfile.lock.  We use this index to compare to the current
source.specs index from the filesystem.

Ironically, this issue was masked by the bug from our prior fix, namely that
dependencies_for_source_changed? always would conclude that there were changes.
Because of that bug, it would short circuit out of the nothing_changed? method
and force a re-resolve with path gems, which would properly build the
Gemfile.lock from the prior parsed one.

Undo some test changes needed to get the first fix to pass.