Removes OSX/msysgit strings from git version
When running the original `version` output through
`Gem::Version.create` when the git version is something like

    $ git --version
    git version 1.2.3 (Apple Git-22)

Ruby blows up with the error:

    ArgumentError:  Malformed version number string 1.2.3 (Apple Git-22)

The '(Apple Git-22)' and any additions that are added by `msysgit` (ex:
1.8.3.msysgit.0) can be safely ignored as they are additions to the git
version for those specific versions.

This was affecting running tests on OSX and being able to even bundle
with git gem dependencies on OSX (with this version of the gem).

This change simply regexes those rules out, in addition to removing the
`git version`, so only the version number (containing numbers or `.`s)
will be parsed.