Add an open_timeout to the fetcher connection
When ruby makes an HTTP connection with the Net::HTTP library, but
default, it has no timeout to setup that connection.  It can be set by
changing the `open_timeout` variable on the connection object, but
beyond that, it will attempt a connection unless an error is returned
from the server.  The Net::HTTP::Persistent library included with
bundler also maintains this same interface when dealing with threads,
and eventually passes that same value down to the Net::HTTP lib.

It is possible to get into a state where bundler is currently attempting
to establish a connection indefinitely to rubygems because this timeout
is not in place.  Using the `Fetcher.api_timeout`, which is just pulled
from the `Bundler.settings[:timeout]` value (defaults to 10 sec), we can
also provide an `open_timeout` value and prevent this.

In most cases, and HTTP connection should be established in less than a
second (if not, there is probably a bigger problem), and most uses of
the Fecther's connection are also wrapped in a Bundler::Retry, so if it
fails do to a network hiccup, it will be attempted again.